Refactor: code-cleanup following C# conventions via RIDER IDE
